JAVA GUI Drawing User Defined Shapes
Author: Cori Nicole Bradshaw

Programming Assignment 2 - JAVA

The basic idea of this project was to create a gui which draws either a
circle, rectangle, or right triangle depending on the user input.
The shapes can either be filled in with color or not. The shapes can
also have a border of the same or differing color. The colors vary from
red, gree, yellow, or blue.

After the user has entered the data and selected from the menu, the
shape is drawn on the screen at the user specified x and y coordinates.
Finally, the user has the option to write the information about the
shape to a text file. This includes the shape's area, perimeter, color,
and location within the X/Y axis.

MVC - model, view, controller idea was also used for this project.